fix syllable counting in compound word handling
Note: one of the fixed regressions is related to an old
hidden mistake: using clen instead of blen of the stem
word lengths was indifferent with the original get_syllable(),
because blen == clen at 8-bit encodings, and UTF-8
words were handled by null-termination. Implementing Unicode
support in Hunspell, clen was changed only in
compound_check_morph() to blen accidentally, but not
in compound_check(), resulting problems from the
recent std::string conversion.

Now this commit is a real fix for the regression from the
commit c63c93237e4decdba5544a96093448605ac549c2,
instead of the following bad fix:

commit d06b0c57ae87ee8743f1bf53f80c1f8e364db619
Author: László Németh <laszlo.nemeth@collabora.com>
Date:   Fri Mar 17 15:11:23 2017 +0100

    fix Hungarian compound word handling