Fix chenc and its usages
chenc was changing buffer it was passed to. This caused untrackable
multiple conversions of token and possibly other variables.

Change it to only return converted string

Additionally logic extending dest buffer implicitly assumed that 0
bytes are left in dest buffer. It's not necessarily the case when
converting to UTF-8 as if result would be 2-byte sequence and only 1 byte
is remaining you get E2BIG as well.

This fixes the case of pipe (-a) in UTF-8 with KOI8-R dictionary.