skip empty ph: field and support character stripping
at replacement rule creation.

When the ph: field ends with the character *,
strip last character of the replacement (the correct word)
and last two character of the field (the * and last
character of the matching pattern) in the replacement rule.

For example,

pretty ph:prity*

dictionary item results "prit -> prett" REP replacement
rule instead of "prity -> pretty", to get
"prity -> pretty" and "pritiest -> prettiest" suggestions.