remove ambiguities from tests/forbiddenword.dic
It's not clear what's meant by having so many entries for `foo`, two of them identical and others ambiguous. Swapping the two Y-flagged entries made the test fail, while the documentation implies no dependency on dictionary order. The behavior may remain undefined in such cases, but let's at least not depend on it in tests.