Keep only REP, ph: or 2-word dictionary phrase suggestions
These are the best suggestions, no need to search other
ones to avoid annoying redundant and long list.

For example to suggest only "a lot" to the bad form "alot",
add the 2-word phrase "a lot" to the dic file.

Or for a very typical spelling mistake, enough to specify the
bad form with a ph: in the dictionary file to remove the other
suggestions.

Note: partial revert of commit de9fe28008eb0761c33bd83847f282602c599fda
"fix up some warnings seen with -Wall -Wextra".