test.sh: remove hardcode libtool path, derive path from Makefile
When running the tests when hunspell was built with slibtoolize +
slibtool most tests fail because of a hardcoded libtool path.

../tests/test.sh: line 93: ../libtool: No such file or directory

This patch removes the hardcoded libtool, derives the value from the
Makefile and sets it as a bash array in test.sh so that the default GNU
libtool value works.

Gentoo-Issue: https://bugs.gentoo.org/958376